mir: Add a value numbering pass
This pass walks the block and assigns value numbers to each member
sequentially. There are probably some interesting issues since this
doesn't cross basic block boundaries, although that it also saving us,
since we don't have to worry about dominance issues.

While this does assign numbers at assignment time, it still needs to
mark the version used.

Fixes: #29

+++ src/mir/mir.hpp
@@ -44,7 +44,7 @@ class Variable {
     std::string name;
 
     /// The version as used by value numbering, 0 means unset
-    uint version;
+    uint64_t version;
 };
 
 class Executable {
